Split level and levelstate files
